% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_metrics.R
\name{composite_metrics}
\alias{composite_metrics}
\title{Calculate Composite Score Metrics (Weights & Loadings)}
\usage{
composite_metrics(
  data = .,
  varlist,
  weight = "correlation",
  metric = c("weights", "loadings")
)
}
\arguments{
\item{data}{A dataframe object. This should be a structured dataset where
each column represents a variable and each row represents an observation.}

\item{varlist}{A required named list of vectors. Each name in the list
represents a composite variable, and the corresponding vector contains the
column names that are associated with the indicators comprising said
composite variable.}

\item{weight}{The weight type to be used in calculating the composite score. 
For composite scores emulating reflective measurements, specify as
`"correlation"` to calculate the composite score with correlation weights.
For composite scores emulating formative measurements, specify as
`"regression"` to calculate the composite score with regression weights.
For composite scores that are unweighted, specify as `"average"` to
calculate the composite scores with averages. The default is set to
correlation weights.}

\item{metric}{The metric (i.e., weights and loadings) to be returned. For
just indicator weights on the composite score, specify as `"weights"`. For
just indicator loadings on the composite score, specify as `"loadings"`.
The default is to return both using `c("weights", "loadings")`.}
}
\value{
A dataframe table with the indicator metrics for their respective
composite.
}
\description{
Create composite score weights and loadings of scales by specifying the
indicators that go into each respective composite variable. Refer to the
help documents for `?regression_loadings`, `?correlation_loadings`, and
`?average_loadings` to see how the loadings are calculated for each
weighting scheme. Refer to the help documents for `?regression_weights`,
`?correlation_weights`, and `?average_weights` to see how the weights are
calculated for each weighting scheme.
}
\examples{

data(grit)

# Specify the named list with composite names and their respective indicators
varlist <- list(extraversion = sprintf("e\%01d", seq(10)),
                neuroticism = sprintf("n\%01d", seq(10)),
                agreeableness = sprintf("a\%01d", seq(10)),
                conscientiousness = sprintf("c\%01d", seq(10)),
                openness = sprintf("o\%01d", seq(10)),
                grit = sprintf("gs\%01d", seq(12)))

# Calculate correlation-weighted composite scores
correlation_metrics <- composite_metrics(data = grit,
                                         varlist = varlist,
                                         weight = "correlation",
                                         metric = c("weights", "loadings"))

# Calculate regression-weighted composite scores
regression_metrics <- composite_metrics(data = grit,
                                        varlist = varlist,
                                        weight = "regression",
                                        metric = c("weights", "loadings"))
                                   
# Calculate unweighted average composite scores
average_metrics <- composite_metrics(data = grit,
                                     varlist = varlist,
                                     weight = "average",
                                     metric = c("weights", "loadings"))

}
