% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_score.R
\name{composite_score}
\alias{composite_score}
\title{Composite Scoring}
\usage{
composite_score(data = ., varlist, weight = "correlation")
}
\arguments{
\item{data}{A dataframe object. This should be a structured dataset where
each column represents a variable and each row represents an observation.}

\item{varlist}{A required named list of vectors. Each name in the list
represents a composite variable, and the corresponding vector contains the
column names that are associated with the indicators comprising said
composite variable.}

\item{weight}{The weight type to be used in calculating the composite score. 
For composite scores emulating reflective measurements, specify as
`"correlation"` to calculate the composite score with correlation weights.
For composite scores emulating formative measurements, specify as
`"regression"` to calculate the composite score with regression weights.
The default is set to correlation weights.}
}
\value{
A dataframe identical to the input dataframe, with additional columns 
appended at the end. These new columns represent the calculated composite 
scores.
}
\description{
Create composite scores of scales by specifying the indicators that go into
each respective composite variable.
}
\examples{

dat <- data(grit)

# Specify the named list with composite names and their respective indicators
varlist <- list(extraversion = sprintf("e\%01d", seq(10)),
                neuroticism = sprintf("n\%01d", seq(10)),
                agreeableness = sprintf("a\%01d", seq(10)),
                conscientiousness = sprintf("c\%01d", seq(10)),
                openness = sprintf("o\%01d", seq(10)),
                grit = sprintf("gs\%01d", seq(12)))

# Calculate correlation weighted composite scores
correlation_data <- composite_score(data = dat,
                                    varlist = varlist,
                                    weight = "correlation")

# Calculate regression weighted composite scores
regression_data <- composite_score(data = dat,
                                   varlist = varlist,
                                   weight = "regression")

}
