% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_weights.R
\name{correlation_weights}
\alias{correlation_weights}
\title{Calculate Correlation-Weighted Composite Score Weights}
\usage{
correlation_weights(data = ., varlist)
}
\arguments{
\item{data}{A dataframe object. This should be a structured dataset where
each column represents a variable and each row represents an observation.}

\item{varlist}{A required named list of vectors. Each name in the list
represents a composite variable, and the corresponding vector contains the
column names that are associated with the indicators comprising said
composite variable.}
}
\value{
A dataframe table with the indicator weights for their respective
composite.
}
\description{
Create composite score weights of scales by specifying the indicators that
go into each respective composite variable. For correlation-weighted
composite scores, weights are calculated as the average correlation of each
indicator with all other indicators.
}
\examples{

data(grit)

# Specify the named list with composite names and their respective indicators
varlist <- list(extraversion = sprintf("e\%01d", seq(10)),
                neuroticism = sprintf("n\%01d", seq(10)),
                agreeableness = sprintf("a\%01d", seq(10)),
                conscientiousness = sprintf("c\%01d", seq(10)),
                openness = sprintf("o\%01d", seq(10)),
                grit = sprintf("gs\%01d", seq(12)))

# Calculate correlation-weighted composite scores
correlation_weights <- correlation_weights(data = grit,
                                           varlist = varlist)

}
