% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_loadings.R
\name{regression_loadings}
\alias{regression_loadings}
\title{Calculate Regression-Weighted Composite Score Loadings}
\usage{
regression_loadings(data = ., varlist)
}
\arguments{
\item{data}{A dataframe object. This should be a structured dataset where
each column represents a variable and each row represents an observation.}

\item{varlist}{A required named list of vectors. Each name in the list
represents a composite variable, and the corresponding vector contains the
column names that are associated with the indicators comprising said
composite variable.}
}
\value{
A dataframe table with the indicator loadings for their respective
composite.
}
\description{
Create composite score loadings of scales by specifying the indicators that
go into each respective composite variable. For regression-weighted composite
scores, loadings are calculated as the bivariate correlation between the
indicator and the regression-weighted composite score.
}
\examples{

data(grit)

# Specify the named list with composite names and their respective indicators
varlist <- list(extraversion = sprintf("e\%01d", seq(10)),
                neuroticism = sprintf("n\%01d", seq(10)),
                agreeableness = sprintf("a\%01d", seq(10)),
                conscientiousness = sprintf("c\%01d", seq(10)),
                openness = sprintf("o\%01d", seq(10)),
                grit = sprintf("gs\%01d", seq(12)))

# Calculate correlation-weighted composite scores & metrics
regression_loadings <- regression_loadings(data = grit,
                                           varlist = varlist)

}
